mainController 분리 fixed #73
